# src/main/resources/META-INF/resources/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "ECskat API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/ecskat/player/cutposition/{pos}" : {
      "put" : {
        "parameters" : [ {
          "name" : "pos",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CutPosition"
          }
        }, {
          "name" : "userId",
          "in" : "cookie",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/ecskat/player/{name}" : {
      "put" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "cookie",
          "schema" : {
            "$ref" : "#/components/schemas/Cookie"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Blatt" : {
        "enum" : [ "ACHT", "ASS", "KOENIG", "NEUN", "OBER", "SIEBEN", "UNTER", "ZEHN" ],
        "type" : "string"
      },
      "Card" : {
        "type" : "object",
        "properties" : {
          "blatt" : {
            "$ref" : "#/components/schemas/Blatt"
          },
          "farbe" : {
            "$ref" : "#/components/schemas/Farbe"
          }
        }
      },
      "Cookie" : {
        "type" : "object",
        "properties" : {
          "domain" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "CutPosition" : {
        "enum" : [ "BOTTOM", "MIDDLE", "TOP" ],
        "type" : "string"
      },
      "Farbe" : {
        "enum" : [ "EICHEL", "GRUEN", "HERZ", "SCHELL" ],
        "type" : "string"
      },
...
      "Player" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "achtzehnGesagt" : {
            "type" : "boolean"
          },
          "cutPosition" : {
            "$ref" : "#/components/schemas/CutPosition"
          },
          "gewonneneStiche" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Card"
            }
          },
          "name" : {
            "type" : "string"
          },
          "ready" : {
            "type" : "boolean"
          }
        }
      },
      "PlayerResponse" : {
        "type" : "object",
        "properties" : {
          "players" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Player"
            }
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}